undo making sort files, add bubble sort ideas to deck files
