--- lib/sort.rb
@@ -1,19 +1,0 @@
-require './lib/card'
-require './lib/guess'
-require './lib/deck'
-require './lib/sort'
-require 'pry'
-
-class Sort
-  attr_accessor
-
-  # def initialize(array)
-  #
-  #
-  #
-  # end
-  #
-  # def bubble_sort
-  #
-  # end
-end
